refactor: move SlugSpec enum to its own module for better organization
