+++ src/cmd/tui.rs
@@ -1,5 +1,5 @@
 use crate::cmd::prs::{self, MergeStateStatus};
-use crate::{graphql, rest};
+use crate::{graphql, rest, slug::SlugSpec};
 use crossterm::{
     event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode, MouseEventKind},
     execute,
@@ -35,28 +35,6 @@ impl MergeStateStatus {
             Self::HasHooks => Color::Yellow,
             Self::Unknown => Color::Magenta,
             Self::Unstable => Color::Yellow,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-enum SlugSpec {
-    Owner(String),
-    Repo { owner: String, name: String },
-}
-
-impl From<&str> for SlugSpec {
-    fn from(s: &str) -> Self {
-        let parts: Vec<&str> = s.split('/').collect();
-        if parts.len() == 1 {
-            SlugSpec::Owner(parts[0].to_string())
-        } else if parts.len() == 2 {
-            SlugSpec::Repo {
-                owner: parts[0].to_string(),
-                name: parts[1].to_string(),
-            }
-        } else {
-            panic!("Invalid slug spec: {}", s);
         }
     }
 }
